Revert "Restructuring to match idiomatic go layout"
This reverts commit 4a9a8caa9d7d11108a7dbc86776d4c0a311f97df.

I really don't understand go package layout. Sorry.
