+++ bot.go
