+++ cmd_test.go
