+++ irc/irc_test.go
